split up the settings screen. each card on that screen now has it's own file
