+++ app/src/main/java/dev/harrisonsoftware/stitchCounter/feature/settings/ExpandableSection.kt
@@ -1,0 +1,107 @@
+package dev.harrisonsoftware.stitchCounter.feature.settings
+
+import androidx.compose.animation.AnimatedVisibility
+import androidx.compose.animation.expandVertically
+import androidx.compose.animation.shrinkVertically
+import androidx.compose.foundation.clickable
+import androidx.compose.foundation.gestures.scrollBy
+import androidx.compose.foundation.layout.Arrangement
+import androidx.compose.foundation.layout.Column
+import androidx.compose.foundation.layout.Row
+import androidx.compose.foundation.layout.fillMaxWidth
+import androidx.compose.foundation.layout.padding
+import androidx.compose.foundation.lazy.LazyListState
+import androidx.compose.material.icons.Icons
+import androidx.compose.material.icons.filled.ExpandLess
+import androidx.compose.material.icons.filled.ExpandMore
+import androidx.compose.material3.Card
+import androidx.compose.material3.CardDefaults
+import androidx.compose.material3.Icon
+import androidx.compose.material3.MaterialTheme
+import androidx.compose.material3.Text
+import androidx.compose.runtime.Composable
+import androidx.compose.runtime.LaunchedEffect
+import androidx.compose.runtime.snapshotFlow
+import androidx.compose.ui.Alignment
+import androidx.compose.ui.Modifier
+import androidx.compose.ui.res.stringResource
+import androidx.compose.ui.semantics.Role
+import androidx.compose.ui.semantics.heading
+import androidx.compose.ui.semantics.semantics
+import androidx.compose.ui.text.font.FontWeight
+import androidx.compose.ui.unit.dp
+import dev.harrisonsoftware.stitchCounter.R
+import kotlinx.coroutines.flow.takeWhile
+
+@Composable
+internal fun ScrollToRevealExpandedItem(
+    isExpanded: Boolean,
+    itemIndex: Int,
+    lazyListState: LazyListState
+) {
+    LaunchedEffect(isExpanded) {
+        if (isExpanded) {
+            val trackingStartTime = System.currentTimeMillis()
+            snapshotFlow { lazyListState.layoutInfo }
+                .takeWhile { System.currentTimeMillis() - trackingStartTime < 500L }
+                .collect { layoutInfo ->
+                    val itemInfo = layoutInfo.visibleItemsInfo.find { it.index == itemIndex }
+                    if (itemInfo != null) {
+                        val overflowBeyondViewport = (itemInfo.offset + itemInfo.size) - layoutInfo.viewportEndOffset
+                        if (overflowBeyondViewport > 0) {
+                            lazyListState.scrollBy(overflowBeyondViewport.toFloat())
+                        }
+                    }
+                }
+        }
+    }
+}
+
+@Composable
+internal fun ExpandableSection(
+    title: String,
+    isExpanded: Boolean,
+    onToggleExpanded: () -> Unit,
+    content: @Composable () -> Unit
+) {
+    Card(
+        modifier = Modifier.fillMaxWidth(),
+        elevation = CardDefaults.cardElevation(defaultElevation = 2.dp),
+        colors = CardDefaults.cardColors().copy(containerColor = MaterialTheme.colorScheme.tertiaryContainer)
+    ) {
+        Column {
+            Row(
+                modifier = Modifier
+                    .fillMaxWidth()
+                    .clickable(role = Role.Button) { onToggleExpanded() }
+                    .padding(16.dp),
+                horizontalArrangement = Arrangement.SpaceBetween,
+                verticalAlignment = Alignment.CenterVertically
+            ) {
+                Text(
+                    text = title,
+                    style = MaterialTheme.typography.headlineMedium,
+                    fontWeight = FontWeight.Bold,
+                    color = MaterialTheme.colorScheme.onTertiaryContainer,
+                    modifier = Modifier.semantics { heading() }
+                )
+                Icon(
+                    imageVector = if (isExpanded) Icons.Default.ExpandLess else Icons.Default.ExpandMore,
+                    contentDescription = if (isExpanded) stringResource(R.string.cd_collapse) else stringResource(R.string.cd_expand)
+                )
+            }
+            
+            AnimatedVisibility(
+                visible = isExpanded,
+                enter = expandVertically(),
+                exit = shrinkVertically()
+            ) {
+                Column(
+                    modifier = Modifier.padding(start = 16.dp, end = 16.dp, bottom = 16.dp, top = 8.dp)
+                ) {
+                    content()
+                }
+            }
+        }
+    }
+}
