+++ app/src/main/java/dev/harrisonsoftware/stitchCounter/feature/settings/SettingsBackupComponents.kt
@@ -1,0 +1,134 @@
+package dev.harrisonsoftware.stitchCounter.feature.settings
+
+import androidx.compose.foundation.layout.Arrangement
+import androidx.compose.foundation.layout.Column
+import androidx.compose.foundation.layout.Spacer
+import androidx.compose.foundation.layout.fillMaxWidth
+import androidx.compose.foundation.layout.padding
+import androidx.compose.foundation.layout.size
+import androidx.compose.foundation.layout.width
+import androidx.compose.material.icons.Icons
+import androidx.compose.material.icons.filled.CloudDownload
+import androidx.compose.material.icons.filled.CloudUpload
+import androidx.compose.material3.AlertDialog
+import androidx.compose.material3.Button
+import androidx.compose.material3.ButtonDefaults
+import androidx.compose.material3.Card
+import androidx.compose.material3.CardDefaults
+import androidx.compose.material3.Icon
+import androidx.compose.material3.MaterialTheme
+import androidx.compose.material3.Text
+import androidx.compose.material3.TextButton
+import androidx.compose.runtime.Composable
+import androidx.compose.ui.Modifier
+import androidx.compose.ui.res.stringResource
+import androidx.compose.ui.unit.dp
+import dev.harrisonsoftware.stitchCounter.R
+import dev.harrisonsoftware.stitchCounter.domain.usecase.ImportResult
+
+@Composable
+internal fun BackupRestoreCard(
+    isExporting: Boolean,
+    isImporting: Boolean,
+    exportError: String?,
+    importError: String?,
+    onExport: () -> Unit,
+    onImport: () -> Unit
+) {
+    Card(
+        modifier = Modifier.fillMaxWidth(),
+        elevation = CardDefaults.cardElevation(defaultElevation = 2.dp),
+        colors = CardDefaults.cardColors().copy(containerColor = MaterialTheme.colorScheme.primaryContainer)
+    ) {
+        Column(
+            modifier = Modifier.padding(16.dp),
+            verticalArrangement = Arrangement.spacedBy(12.dp)
+        ) {
+            Button(
+                onClick = onExport,
+                enabled = !isExporting && !isImporting,
+                modifier = Modifier.fillMaxWidth()
+            ) {
+                Icon(
+                    imageVector = Icons.Default.CloudUpload,
+                    contentDescription = null,
+                    modifier = Modifier.size(18.dp)
+                )
+                Spacer(modifier = Modifier.width(8.dp))
+                Text(
+                    if (isExporting) stringResource(R.string.settings_exporting)
+                    else stringResource(R.string.settings_export_library)
+                )
+            }
+            
+            if (exportError != null) {
+                Text(
+                    text = stringResource(R.string.settings_error_format, exportError),
+                    color = MaterialTheme.colorScheme.error,
+                    style = MaterialTheme.typography.bodySmall
+                )
+            }
+            
+            Button(
+                onClick = onImport,
+                enabled = !isExporting && !isImporting,
+                modifier = Modifier.fillMaxWidth(),
+                colors = ButtonDefaults.buttonColors(
+                    containerColor = MaterialTheme.colorScheme.secondary
+                )
+            ) {
+                Icon(
+                    imageVector = Icons.Default.CloudDownload,
+                    contentDescription = null,
+                    modifier = Modifier.size(18.dp)
+                )
+                Spacer(modifier = Modifier.width(8.dp))
+                Text(
+                    if (isImporting) stringResource(R.string.settings_importing)
+                    else stringResource(R.string.settings_import_library)
+                )
+            }
+            
+            if (importError != null) {
+                Text(
+                    text = stringResource(R.string.settings_error_format, importError),
+                    color = MaterialTheme.colorScheme.error,
+                    style = MaterialTheme.typography.bodySmall
+                )
+            }
+        }
+    }
+}
+
+@Composable
+internal fun ImportResultDialog(
+    result: ImportResult,
+    onDismiss: () -> Unit
+) {
+    AlertDialog(
+        onDismissRequest = onDismiss,
+        title = { Text(stringResource(R.string.settings_import_complete)) },
+        text = {
+            Column(verticalArrangement = Arrangement.spacedBy(8.dp)) {
+                Text(stringResource(R.string.settings_imported_count, result.importedCount))
+                if (result.failedCount > 0) {
+                    Text(
+                        stringResource(R.string.settings_failed_import_count, result.failedCount),
+                        color = MaterialTheme.colorScheme.error
+                    )
+                    if (result.failedProjectNames.isNotEmpty()) {
+                        Text(
+                            result.failedProjectNames.joinToString("\n"),
+                            style = MaterialTheme.typography.bodySmall
+                        )
+                    }
+                }
+            }
+        },
+        confirmButton = {
+            TextButton(onClick = onDismiss) {
+                Text(stringResource(R.string.action_ok))
+            }
+        }
+    )
+}
