+++ app/src/main/java/dev/harrisonsoftware/stitchCounter/feature/settings/SettingsScreen.kt
@@ -4,57 +4,26 @@ import android.content.Intent
 import android.net.Uri
 import androidx.activity.compose.rememberLauncherForActivityResult
 import androidx.activity.result.contract.ActivityResultContracts
-import androidx.compose.foundation.background
-import androidx.compose.foundation.clickable
 import androidx.compose.foundation.layout.*
-import androidx.compose.foundation.selection.selectable
 import androidx.compose.foundation.lazy.LazyColumn
 import androidx.compose.foundation.lazy.rememberLazyListState
-import androidx.compose.foundation.shape.CircleShape
-import androidx.compose.animation.AnimatedVisibility
-import androidx.compose.animation.expandVertically
-import androidx.compose.animation.shrinkVertically
-import androidx.compose.foundation.gestures.scrollBy
-import androidx.compose.material.icons.Icons
-import androidx.compose.material.icons.filled.BugReport
-import androidx.compose.material.icons.filled.CloudDownload
-import androidx.compose.material.icons.filled.CloudUpload
-import androidx.compose.material.icons.filled.ExpandLess
-import androidx.compose.material.icons.filled.ExpandMore
-import androidx.compose.material.icons.filled.Feedback
-import androidx.compose.material.icons.filled.Lightbulb
-import androidx.compose.material.icons.filled.LocalPolice
-import androidx.compose.material.icons.filled.Policy
 import androidx.compose.material3.*
-import androidx.compose.material3.CardDefaults
-import androidx.compose.ui.semantics.Role
-import androidx.compose.ui.semantics.contentDescription
-import androidx.compose.ui.semantics.heading
-import androidx.compose.ui.semantics.semantics
-import androidx.compose.foundation.lazy.LazyListState
 import androidx.compose.runtime.Composable
 import androidx.compose.runtime.LaunchedEffect
 import androidx.compose.runtime.getValue
 import androidx.compose.runtime.mutableStateOf
 import androidx.compose.runtime.remember
-import androidx.compose.runtime.snapshotFlow
-import androidx.compose.ui.Alignment
 import androidx.compose.ui.Modifier
-import androidx.compose.ui.draw.clip
 import androidx.compose.ui.platform.LocalContext
 import androidx.compose.ui.res.stringResource
-import androidx.compose.ui.text.font.FontWeight
 import androidx.compose.ui.unit.dp
 import androidx.hilt.navigation.compose.hiltViewModel
 import androidx.lifecycle.compose.collectAsStateWithLifecycle
 import dev.harrisonsoftware.stitchCounter.R
 import dev.harrisonsoftware.stitchCounter.domain.model.AppTheme
 import dev.harrisonsoftware.stitchCounter.feature.navigation.RootNavGraph
-import dev.harrisonsoftware.stitchCounter.feature.sharedComposables.ThemedAppIcon
-import dev.harrisonsoftware.stitchCounter.feature.theme.ThemeColor
 import com.ramcosta.composedestinations.annotation.Destination
 import androidx.core.net.toUri
-import kotlinx.coroutines.flow.takeWhile
 import dev.harrisonsoftware.stitchCounter.Constants
 
 @RootNavGraph
@@ -237,416 +206,4 @@ fun SettingsScreen(
             }
         }
     }
-}
-
-@Composable
-private fun ScrollToRevealExpandedItem(
-    isExpanded: Boolean,
-    itemIndex: Int,
-    lazyListState: LazyListState
-) {
-    LaunchedEffect(isExpanded) {
-        if (isExpanded) {
-            val trackingStartTime = System.currentTimeMillis()
-            snapshotFlow { lazyListState.layoutInfo }
-                .takeWhile { System.currentTimeMillis() - trackingStartTime < 500L }
-                .collect { layoutInfo ->
-                    val itemInfo = layoutInfo.visibleItemsInfo.find { it.index == itemIndex }
-                    if (itemInfo != null) {
-                        val overflowBeyondViewport = (itemInfo.offset + itemInfo.size) - layoutInfo.viewportEndOffset
-                        if (overflowBeyondViewport > 0) {
-                            lazyListState.scrollBy(overflowBeyondViewport.toFloat())
-                        }
-                    }
-                }
-        }
-    }
-}
-
-@Composable
-private fun ExpandableSection(
-    title: String,
-    isExpanded: Boolean,
-    onToggleExpanded: () -> Unit,
-    content: @Composable () -> Unit
-) {
-    Card(
-        modifier = Modifier.fillMaxWidth(),
-        elevation = CardDefaults.cardElevation(defaultElevation = 2.dp),
-        colors = CardDefaults.cardColors().copy(containerColor = MaterialTheme.colorScheme.tertiaryContainer)
-    ) {
-        Column {
-            Row(
-                modifier = Modifier
-                    .fillMaxWidth()
-                    .clickable(role = Role.Button) { onToggleExpanded() }
-                    .padding(16.dp),
-                horizontalArrangement = Arrangement.SpaceBetween,
-                verticalAlignment = Alignment.CenterVertically
-            ) {
-                Text(
-                    text = title,
-                    style = MaterialTheme.typography.headlineMedium,
-                    fontWeight = FontWeight.Bold,
-                    color = MaterialTheme.colorScheme.onTertiaryContainer,
-                    modifier = Modifier.semantics { heading() }
-                )
-                Icon(
-                    imageVector = if (isExpanded) Icons.Default.ExpandLess else Icons.Default.ExpandMore,
-                    contentDescription = if (isExpanded) stringResource(R.string.cd_collapse) else stringResource(R.string.cd_expand)
-                )
-            }
-            
-            AnimatedVisibility(
-                visible = isExpanded,
-                enter = expandVertically(),
-                exit = shrinkVertically()
-            ) {
-                Column(
-                    modifier = Modifier.padding(start = 16.dp, end = 16.dp, bottom = 16.dp, top = 8.dp)
-                ) {
-                    content()
-                }
-            }
-        }
-    }
-}
-
-@Composable
-private fun ThemeOptionCard(
-    theme: AppTheme,
-    isSelected: Boolean,
-    themeColors: List<ThemeColor>,
-    onThemeSelected: () -> Unit
-) {
-    Card(
-        modifier = Modifier
-            .fillMaxWidth()
-            .selectable(
-                selected = isSelected,
-                role = Role.RadioButton,
-                onClick = onThemeSelected
-            ),
-        colors = CardDefaults.cardColors(
-            containerColor = if (isSelected) MaterialTheme.colorScheme.primaryContainer
-            else MaterialTheme.colorScheme.surface
-        ),
-        elevation = CardDefaults.cardElevation(
-            defaultElevation = if (isSelected) 8.dp else 2.dp
-        )
-    ) {
-        Column(
-            modifier = Modifier.padding(16.dp)
-        ) {
-            Row(
-                modifier = Modifier.fillMaxWidth(),
-                horizontalArrangement = Arrangement.SpaceBetween,
-                verticalAlignment = Alignment.CenterVertically
-            ) {
-                Row(
-                    verticalAlignment = Alignment.CenterVertically,
-                    horizontalArrangement = Arrangement.spacedBy(12.dp)
-                ) {
-                    ThemedAppIcon(
-                        theme = theme,
-                        modifier = Modifier.size(48.dp),
-                        isSelected = isSelected
-                    )
-                    Text(
-                        text = theme.displayName,
-                        style = MaterialTheme.typography.titleLarge,
-                        fontWeight = FontWeight.Bold
-                    )
-                }
-
-                RadioButton(
-                    selected = isSelected,
-                    onClick = onThemeSelected
-                )
-            }
-            
-            if (isSelected && themeColors.isNotEmpty()) {
-                Spacer(modifier = Modifier.height(12.dp))
-                Text(
-                    text = stringResource(R.string.settings_colors_in_theme),
-                    style = MaterialTheme.typography.bodyMedium,
-                    color = MaterialTheme.colorScheme.onSurfaceVariant
-                )
-                
-                Spacer(modifier = Modifier.height(8.dp))
-                
-                Column(
-                    verticalArrangement = Arrangement.spacedBy(8.dp)
-                ) {
-                    themeColors.forEach { themeColor ->
-                        ColorItem(themeColor = themeColor)
-                    }
-                }
-            }
-        }
-    }
-}
-
-@Composable
-private fun ColorItem(themeColor: ThemeColor) {
-    val lightSwatchDescription = stringResource(R.string.cd_color_swatch_light, themeColor.name)
-    val darkSwatchDescription = stringResource(R.string.cd_color_swatch_dark, themeColor.name)
-
-    Row(
-        modifier = Modifier.fillMaxWidth(),
-        verticalAlignment = Alignment.CenterVertically,
-        horizontalArrangement = Arrangement.spacedBy(12.dp)
-    ) {
-        Box(
-            modifier = Modifier
-                .size(24.dp)
-                .clip(CircleShape)
-                .background(themeColor.lightColor)
-                .semantics { contentDescription = lightSwatchDescription }
-        )
-        
-        Text(
-            text = stringResource(R.string.settings_color_light_format, themeColor.name),
-            style = MaterialTheme.typography.bodySmall,
-            color = MaterialTheme.colorScheme.onSurfaceVariant
-        )
-        
-        Spacer(modifier = Modifier.weight(1f))
-        
-        Box(
-            modifier = Modifier
-                .size(24.dp)
-                .clip(CircleShape)
-                .background(themeColor.darkColor)
-                .semantics { contentDescription = darkSwatchDescription }
-        )
-        
-        Text(
-            text = stringResource(R.string.settings_dark),
-            style = MaterialTheme.typography.bodySmall,
-            color = MaterialTheme.colorScheme.onSurfaceVariant
-        )
-    }
-}
-
-@Composable
-private fun BackupRestoreCard(
-    isExporting: Boolean,
-    isImporting: Boolean,
-    exportError: String?,
-    importError: String?,
-    onExport: () -> Unit,
-    onImport: () -> Unit
-) {
-    Card(
-        modifier = Modifier.fillMaxWidth(),
-        elevation = CardDefaults.cardElevation(defaultElevation = 2.dp),
-        colors = CardDefaults.cardColors().copy(containerColor = MaterialTheme.colorScheme.primaryContainer)
-    ) {
-        Column(
-            modifier = Modifier.padding(16.dp),
-            verticalArrangement = Arrangement.spacedBy(12.dp)
-        ) {
-            Button(
-                onClick = onExport,
-                enabled = !isExporting && !isImporting,
-                modifier = Modifier.fillMaxWidth()
-            ) {
-                Icon(
-                    imageVector = Icons.Default.CloudUpload,
-                    contentDescription = null,
-                    modifier = Modifier.size(18.dp)
-                )
-                Spacer(modifier = Modifier.width(8.dp))
-                Text(
-                    if (isExporting) stringResource(R.string.settings_exporting)
-                    else stringResource(R.string.settings_export_library)
-                )
-            }
-            
-            if (exportError != null) {
-                Text(
-                    text = stringResource(R.string.settings_error_format, exportError),
-                    color = MaterialTheme.colorScheme.error,
-                    style = MaterialTheme.typography.bodySmall
-                )
-            }
-            
-            Button(
-                onClick = onImport,
-                enabled = !isExporting && !isImporting,
-                modifier = Modifier.fillMaxWidth(),
-                colors = ButtonDefaults.buttonColors(
-                    containerColor = MaterialTheme.colorScheme.secondary
-                )
-            ) {
-                Icon(
-                    imageVector = Icons.Default.CloudDownload,
-                    contentDescription = null,
-                    modifier = Modifier.size(18.dp)
-                )
-                Spacer(modifier = Modifier.width(8.dp))
-                Text(
-                    if (isImporting) stringResource(R.string.settings_importing)
-                    else stringResource(R.string.settings_import_library)
-                )
-            }
-            
-            if (importError != null) {
-                Text(
-                    text = stringResource(R.string.settings_error_format, importError),
-                    color = MaterialTheme.colorScheme.error,
-                    style = MaterialTheme.typography.bodySmall
-                )
-            }
-        }
-    }
-}
-
-@Composable
-private fun SupportCard(
-    onReportBug: () -> Unit,
-    onGiveFeedback: () -> Unit,
-    onRequestFeature: () -> Unit
-) {
-    Card(
-        modifier = Modifier.fillMaxWidth(),
-        elevation = CardDefaults.cardElevation(defaultElevation = 2.dp),
-        colors = CardDefaults.cardColors().copy(containerColor = MaterialTheme.colorScheme.primaryContainer)
-    ) {
-        Column(
-            modifier = Modifier.padding(16.dp),
-            verticalArrangement = Arrangement.spacedBy(12.dp)
-        ) {
-            Button(
-                onClick = onReportBug,
-                modifier = Modifier.fillMaxWidth()
-            ) {
-                Icon(
-                    imageVector = Icons.Default.BugReport,
-                    contentDescription = null,
-                    modifier = Modifier.size(18.dp)
-                )
-                Spacer(modifier = Modifier.width(8.dp))
-                Text(stringResource(R.string.settings_report_bug))
-            }
-
-            Button(
-                onClick = onGiveFeedback,
-                modifier = Modifier.fillMaxWidth(),
-                colors = ButtonDefaults.buttonColors(
-                    containerColor = MaterialTheme.colorScheme.secondary
-                )
-            ) {
-                Icon(
-                    imageVector = Icons.Default.Feedback,
-                    contentDescription = null,
-                    modifier = Modifier.size(18.dp)
-                )
-                Spacer(modifier = Modifier.width(8.dp))
-                Text(stringResource(R.string.settings_give_feedback))
-            }
-
-            Button(
-                onClick = onRequestFeature,
-                modifier = Modifier.fillMaxWidth(),
-                colors = ButtonDefaults.buttonColors(
-                    containerColor = MaterialTheme.colorScheme.tertiary
-                )
-            ) {
-                Icon(
-                    imageVector = Icons.Default.Lightbulb,
-                    contentDescription = null,
-                    modifier = Modifier.size(18.dp)
-                )
-                Spacer(modifier = Modifier.width(8.dp))
-                Text(stringResource(R.string.settings_request_feature))
-            }
-        }
-    }
-}
-
-@Composable
-private fun LegalCard(
-    onOpenPrivacyPolicy: () -> Unit,
-    onOpenEULA: () -> Unit
-) {
-    Card(
-        modifier = Modifier.fillMaxWidth(),
-        elevation = CardDefaults.cardElevation(defaultElevation = 2.dp),
-        colors = CardDefaults.cardColors().copy(containerColor = MaterialTheme.colorScheme.primaryContainer)
-    ) {
-        Column(
-            modifier = Modifier.padding(16.dp),
-            verticalArrangement = Arrangement.spacedBy(12.dp)
-        ) {
-
-            Button(
-                onClick = onOpenPrivacyPolicy,
-                modifier = Modifier.fillMaxWidth(),
-                colors = ButtonDefaults.buttonColors(
-                    containerColor = MaterialTheme.colorScheme.primary
-                )
-            ) {
-                Icon(
-                    imageVector = Icons.Default.Policy,
-                    contentDescription = null,
-                    modifier = Modifier.size(18.dp)
-                )
-                Spacer(modifier = Modifier.width(8.dp))
-                Text(stringResource(R.string.settings_privacy_policy))
-            }
-
-            Button(
-                onClick = onOpenEULA,
-                modifier = Modifier.fillMaxWidth(),
-                colors = ButtonDefaults.buttonColors(
-                    containerColor = MaterialTheme.colorScheme.secondary
-                )
-            ) {
-                Icon(
-                    imageVector = Icons.Default.LocalPolice,
-                    contentDescription = null,
-                    modifier = Modifier.size(18.dp)
-                )
-                Spacer(modifier = Modifier.width(8.dp))
-                Text(stringResource(R.string.settings_eula))
-            }
-
-            Text(stringResource(R.string.settings_eula_description))
-        }
-    }
-}
-
-@Composable
-private fun ImportResultDialog(
-    result: dev.harrisonsoftware.stitchCounter.domain.usecase.ImportResult,
-    onDismiss: () -> Unit
-) {
-    AlertDialog(
-        onDismissRequest = onDismiss,
-        title = { Text(stringResource(R.string.settings_import_complete)) },
-        text = {
-            Column(verticalArrangement = Arrangement.spacedBy(8.dp)) {
-                Text(stringResource(R.string.settings_imported_count, result.importedCount))
-                if (result.failedCount > 0) {
-                    Text(
-                        stringResource(R.string.settings_failed_import_count, result.failedCount),
-                        color = MaterialTheme.colorScheme.error
-                    )
-                    if (result.failedProjectNames.isNotEmpty()) {
-                        Text(
-                            result.failedProjectNames.joinToString("\n"),
-                            style = MaterialTheme.typography.bodySmall
-                        )
-                    }
-                }
-            }
-        },
-        confirmButton = {
-            TextButton(onClick = onDismiss) {
-                Text(stringResource(R.string.action_ok))
-            }
-        }
-    )
 }
