+++ app/src/main/java/dev/harrisonsoftware/stitchCounter/feature/settings/SettingsThemeComponents.kt
@@ -1,0 +1,150 @@
+package dev.harrisonsoftware.stitchCounter.feature.settings
+
+import androidx.compose.foundation.background
+import androidx.compose.foundation.layout.Arrangement
+import androidx.compose.foundation.layout.Box
+import androidx.compose.foundation.layout.Column
+import androidx.compose.foundation.layout.Row
+import androidx.compose.foundation.layout.Spacer
+import androidx.compose.foundation.layout.fillMaxWidth
+import androidx.compose.foundation.layout.height
+import androidx.compose.foundation.layout.padding
+import androidx.compose.foundation.layout.size
+import androidx.compose.foundation.selection.selectable
+import androidx.compose.foundation.shape.CircleShape
+import androidx.compose.material3.Card
+import androidx.compose.material3.CardDefaults
+import androidx.compose.material3.MaterialTheme
+import androidx.compose.material3.RadioButton
+import androidx.compose.material3.Text
+import androidx.compose.runtime.Composable
+import androidx.compose.ui.Alignment
+import androidx.compose.ui.Modifier
+import androidx.compose.ui.draw.clip
+import androidx.compose.ui.res.stringResource
+import androidx.compose.ui.semantics.Role
+import androidx.compose.ui.semantics.contentDescription
+import androidx.compose.ui.semantics.semantics
+import androidx.compose.ui.text.font.FontWeight
+import androidx.compose.ui.unit.dp
+import dev.harrisonsoftware.stitchCounter.R
+import dev.harrisonsoftware.stitchCounter.domain.model.AppTheme
+import dev.harrisonsoftware.stitchCounter.feature.sharedComposables.ThemedAppIcon
+import dev.harrisonsoftware.stitchCounter.feature.theme.ThemeColor
+
+@Composable
+internal fun ThemeOptionCard(
+    theme: AppTheme,
+    isSelected: Boolean,
+    themeColors: List<ThemeColor>,
+    onThemeSelected: () -> Unit
+) {
+    Card(
+        modifier = Modifier
+            .fillMaxWidth()
+            .selectable(
+                selected = isSelected,
+                role = Role.RadioButton,
+                onClick = onThemeSelected
+            ),
+        colors = CardDefaults.cardColors(
+            containerColor = if (isSelected) MaterialTheme.colorScheme.primaryContainer
+            else MaterialTheme.colorScheme.surface
+        ),
+        elevation = CardDefaults.cardElevation(
+            defaultElevation = if (isSelected) 8.dp else 2.dp
+        )
+    ) {
+        Column(
+            modifier = Modifier.padding(16.dp)
+        ) {
+            Row(
+                modifier = Modifier.fillMaxWidth(),
+                horizontalArrangement = Arrangement.SpaceBetween,
+                verticalAlignment = Alignment.CenterVertically
+            ) {
+                Row(
+                    verticalAlignment = Alignment.CenterVertically,
+                    horizontalArrangement = Arrangement.spacedBy(12.dp)
+                ) {
+                    ThemedAppIcon(
+                        theme = theme,
+                        modifier = Modifier.size(48.dp),
+                        isSelected = isSelected
+                    )
+                    Text(
+                        text = theme.displayName,
+                        style = MaterialTheme.typography.titleLarge,
+                        fontWeight = FontWeight.Bold
+                    )
+                }
+
+                RadioButton(
+                    selected = isSelected,
+                    onClick = onThemeSelected
+                )
+            }
+            
+            if (isSelected && themeColors.isNotEmpty()) {
+                Spacer(modifier = Modifier.height(12.dp))
+                Text(
+                    text = stringResource(R.string.settings_colors_in_theme),
+                    style = MaterialTheme.typography.bodyMedium,
+                    color = MaterialTheme.colorScheme.onSurfaceVariant
+                )
+                
+                Spacer(modifier = Modifier.height(8.dp))
+                
+                Column(
+                    verticalArrangement = Arrangement.spacedBy(8.dp)
+                ) {
+                    themeColors.forEach { themeColor ->
+                        ColorItem(themeColor = themeColor)
+                    }
+                }
+            }
+        }
+    }
+}
+
+@Composable
+internal fun ColorItem(themeColor: ThemeColor) {
+    val lightSwatchDescription = stringResource(R.string.cd_color_swatch_light, themeColor.name)
+    val darkSwatchDescription = stringResource(R.string.cd_color_swatch_dark, themeColor.name)
+
+    Row(
+        modifier = Modifier.fillMaxWidth(),
+        verticalAlignment = Alignment.CenterVertically,
+        horizontalArrangement = Arrangement.spacedBy(12.dp)
+    ) {
+        Box(
+            modifier = Modifier
+                .size(24.dp)
+                .clip(CircleShape)
+                .background(themeColor.lightColor)
+                .semantics { contentDescription = lightSwatchDescription }
+        )
+        
+        Text(
+            text = stringResource(R.string.settings_color_light_format, themeColor.name),
+            style = MaterialTheme.typography.bodySmall,
+            color = MaterialTheme.colorScheme.onSurfaceVariant
+        )
+        
+        Spacer(modifier = Modifier.weight(1f))
+        
+        Box(
+            modifier = Modifier
+                .size(24.dp)
+                .clip(CircleShape)
+                .background(themeColor.darkColor)
+                .semantics { contentDescription = darkSwatchDescription }
+        )
+        
+        Text(
+            text = stringResource(R.string.settings_dark),
+            style = MaterialTheme.typography.bodySmall,
+            color = MaterialTheme.colorScheme.onSurfaceVariant
+        )
+    }
+}
